refactor(workflow): consolidate LLM-related imports and update events
- Moved and renamed `LLMNode` from `llm_node` to `node` for clearer structuring.
- Introduced `ModelInvokeCompletedEvent` to replace the `ModelInvokeCompleted` class.
- Updated import paths throughout to reflect module restructuring.
- Updated event handling logic to use `ModelInvokeCompletedEvent` consistently.
