+++ api/core/workflow/nodes/event/__init__.py
@@ -1,4 +1,10 @@
-from .event import RunCompletedEvent, RunRetrieverResourceEvent, RunStreamChunkEvent
+from .event import ModelInvokeCompletedEvent, RunCompletedEvent, RunRetrieverResourceEvent, RunStreamChunkEvent
 from .types import NodeEvent
 
-__all__ = ["RunCompletedEvent", "RunRetrieverResourceEvent", "RunStreamChunkEvent", "NodeEvent"]
+__all__ = [
+    "RunCompletedEvent",
+    "RunRetrieverResourceEvent",
+    "RunStreamChunkEvent",
+    "NodeEvent",
+    "ModelInvokeCompletedEvent",
+]
