+++ api/tests/unit_tests/core/workflow/graph_engine/test_graph_engine.py
@@ -18,7 +18,7 @@ from core.workflow.graph_engine.entities.graph import Graph
 from core.workflow.graph_engine.entities.runtime_route_state import RouteNodeState
 from core.workflow.graph_engine.graph_engine import GraphEngine
 from core.workflow.nodes.event import RunCompletedEvent, RunStreamChunkEvent
-from core.workflow.nodes.llm.llm_node import LLMNode
+from core.workflow.nodes.llm.node import LLMNode
 from enums import UserFrom
 from models.workflow import WorkflowNodeExecutionStatus, WorkflowType
